requirements.txt
pyzmq==17.1.2
pyzmq-static==2.2
